Create count down timer component and place header styles in an object
